+ Incorpore botones para guardar y cargar las preferencias del usuario! :D

+++ Custom Playground/popup.js
@@ -1,10 +1,15 @@
-const selectores = document.querySelectorAll('.colorSelector');
-const deleteStorage = document.querySelector('.resetStorage');
-const resetPg = document.querySelector('.restoreAll');
-const restoreOne = document.querySelectorAll('.restore')
+// elementos para los tabs
 const ol = document.querySelectorAll('.ol');
 const bloque = document.querySelectorAll('.bloque');
-const darkPg = document.querySelector('.darkPg')
+// los selectores de colores
+const selectores = document.querySelectorAll('.colorSelector');
+// botones
+const restoreOne = document.querySelectorAll('.restore')
+const deleteStorage = document.querySelector('.resetStorage'); // dev-only
+const resetPg = document.querySelector('.restoreAll');
+const saveAll = document.querySelector('.saveAll');
+const loadUserSettings = document.querySelector('.loadUserSettings');
+const darkPg = document.querySelector('.darkPg');
 
 // configuracion de los tabs
 ol.forEach((eachOl, i) => {
@@ -29,15 +34,27 @@ function syncInputValues() {
     chrome.storage.sync.get(function (userSettings) {
         for (const keyName in userSettings) {
             if (Object.hasOwnProperty.call(userSettings, keyName)) {
-                const value = userSettings[keyName];
-                const input = document.querySelector(`#${keyName}`);
-                if (value && input) {
+                if (userSettings.save) {
+                    if (keyName.startsWith('u')) {
+                        const value = userSettings[keyName];
+                        const newName = keyName.slice(1)
+                        const input = document.querySelector(`#${newName}`);
+                        if (value && input) {
+                            input.setAttribute('value', `${value}`)
+                        }
+                    }
+                } else {
+                    const value = userSettings[keyName];
+                    const input = document.querySelector(`#${keyName}`);
+                    if (value && input) {
                     input.setAttribute('value', `${value}`)
+                    }
                 }
             }
         }
     })
 }
+
 
 // escuchar cambios en los selectores de color, adjuntar cada valor y enviar al storage
 selectores.forEach(input => {
@@ -133,7 +150,11 @@ selectores.forEach(input => {
 
 // delete storage? (dev-only)
 deleteStorage.addEventListener('click', function () {
-    chrome.storage.sync.clear();
+    if (confirm("Esto VACIARA por completo el storage de chrome, estas seguro?")) {
+        chrome.storage.sync.clear();
+    } else {
+        deleteStorage.preventDefault()
+    }
 })
 
 // restaurar cada uno
@@ -246,9 +267,109 @@ deleteStorage.addEventListener('click', function () {
     })
 }) 
 
+// guardar settings del usuario!
+saveAll.addEventListener('click', function () {
+    selectores.forEach(input => {
+        if (input.id === 'brand') {
+            chrome.storage.sync.set({
+                'ubrand': input.value,
+            })
+        }
+        if (input.id === 'white') {
+            chrome.storage.sync.set({
+                'uwhite': input.value,
+            })
+        }
+        if (input.id === 'commonText') {
+            chrome.storage.sync.set({
+                'ucommonText': input.value,
+            })
+        }
+        if (input.id === 'disabled') {
+            chrome.storage.sync.set({
+                'udisabled': input.value,
+            })
+        }
+        if (input.id === 'lightGray') {
+            chrome.storage.sync.set({
+                'ulightGray': input.value,
+            })
+        }
+        if (input.id === 'background') {
+            chrome.storage.sync.set({
+                'ubackground': input.value,
+            })
+        }
+        if (input.id === 'primary') {
+            chrome.storage.sync.set({
+                'uprimary': input.value,
+            })
+        }
+        if (input.id === 'primaryDark') {
+            chrome.storage.sync.set({
+                'uprimaryDark': input.value,
+            })
+        }
+        if (input.id === 'primaryLight') {
+            chrome.storage.sync.set({
+                'uprimaryLight': input.value,
+            })
+        }
+        if (input.id === 'success') {
+            chrome.storage.sync.set({
+                'usuccess': input.value,
+            })
+        }
+        if (input.id === 'successLight') {
+            chrome.storage.sync.set({
+                'usuccessLight': input.value,
+            })
+        }
+        if (input.id === 'warning') {
+            chrome.storage.sync.set({
+                'uwarning': input.value,
+            })
+        }
+        if (input.id === 'warningLight') {
+            chrome.storage.sync.set({
+                'uwarningLight': input.value,
+            })
+        }
+        if (input.id === 'info') {
+            chrome.storage.sync.set({
+                'uinfo': input.value,
+            })
+        }
+        if (input.id === 'infoLight') {
+            chrome.storage.sync.set({
+                'uinfoLight': input.value,
+            })
+        }
+        if (input.id === 'danger') {
+            chrome.storage.sync.set({
+                'udanger': input.value,
+            })
+        }
+        if (input.id === 'dangerLight') {
+            chrome.storage.sync.set({
+                'udangerLight': input.value,
+            })
+        }
+    })
+})
+
+// cargar perfil del usuario
+loadUserSettings.addEventListener('click', function () {
+    chrome.storage.sync.set({
+        'save': true,
+    })
+    syncInputValues()
+})
+
 // retornar a la config default de PG?
 resetPg.addEventListener('click', function () {
     chrome.storage.sync.set({
+        'save': false,
         'brand': '#cb1e40',
         'white': '#ffffff',
         'commonText': '#303030',
@@ -273,6 +394,7 @@ resetPg.addEventListener('click', function () {
 // Dark PG settings
 darkPg.addEventListener('click', function () {
     chrome.storage.sync.set({
+        'save': false,
         'brand': '#01b100',
         'white': '#262626',
         'commonText': '#f1f1f1',
